requirements.txt
huggingface_hub
pyyaml
prompt_toolkit